api_gateway/requirements.txt
flask==3.0.3
requests==2.32.3
pyjwt==2.9.0